🚀 Final Sync: Database IDs unified to Long for League, Team, and Match

+++ android-app/app/src/main/java/com/duallive/app/data/entity/Team.kt
@@ -5,9 +5,9 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "teams")
 data class Team(
-    @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val leagueId: Int, // Link to the League
+    @PrimaryKey(autoGenerate = true) val id: Long = 0,
+    val leagueId: Long, 
     val name: String,
     val logoPath: String? = null,
-    val groupName: String? = null // For UCL groups
+    val groupName: String? = null 
 )
